- fix warning - fix occasional crash when switching views (the delayed deletion was the   culprit, Stephan, what exactly was the reason for that at all? I can't   remember)
svn path=/trunk/kdelibs/; revision=60510

+++ kfile/kfiledialog.cpp
@@ -362,7 +362,7 @@ KMimeType::Ptr KFileDialog::currentFilterMimeType()
 {
     int i = filterWidget->currentItem()-1;
 
-    if ((i >= 0) && (i < d->mimetypes.count()))
+    if ((i >= 0) && (i < (int) d->mimetypes.count()))
        return d->mimetypes[i];
     return d->defaultType;
 }
